+++ rumqttc/src/v5/mod.rs
@@ -21,7 +21,8 @@ use crate::{NetworkOptions, Transport};
 use mqttbytes::v5::*;
 
 pub use client::{
-    AsyncClient, Client, ClientError, Connection, Iter, RecvError, RecvTimeoutError, TryRecvError,
+    AsyncClient, Client, ClientError, Connection, InvalidTopic, Iter, RecvError, RecvTimeoutError,
+    Topic, TryRecvError, ValidatedTopic,
 };
 pub use eventloop::{ConnectionError, Event, EventLoop};
 pub use state::{MqttState, StateError};